+++ front/src/pages/products/[productId].tsx
@@ -14,6 +14,8 @@ const Product = () => {
   const idFromUrl = router.query.productId;
   console.log(idFromUrl);
 
+  const tokenFromLS = localStorage.getItem("token");
+
   const [product, setProduct] = useState<Product | null>(null);
 
   useEffect(() => {
@@ -23,7 +25,12 @@ const Product = () => {
       }
       if (localStorage.getItem("token")) {
         const response = await fetch(
-          `http://localhost:3001/products/${idFromUrl}`
+          `http://localhost:3001/products/${idFromUrl}`,
+          {
+            headers: {
+              Authorization: "Bearer " + tokenFromLS,
+            },
+          }
         );
         const data = await response.json();
 
@@ -34,22 +41,27 @@ const Product = () => {
   }, [idFromUrl]);
 
   const handleDelete = async (idFromUrl: number) => {
-    try {
-      const response = await fetch(
-        `http://localhost:3001/products/delete/${idFromUrl}`,
-        {
-          method: "DELETE",
-        }
-      );
+    if (localStorage.getItem("token")) {
+      try {
+        const response = await fetch(
+          `http://localhost:3001/products/delete/${idFromUrl}`,
+          {
+            headers: {
+              Authorization: "Bearer " + tokenFromLS,
+            },
+            method: "DELETE",
+          }
+        );
 
-      if (response.ok) {
-        console.log("Product deleted successfully");
-        router.push("/main");
-      } else {
-        console.error("Error deleting product");
+        if (response.ok) {
+          console.log("Product deleted successfully");
+          router.push("/my-products");
+        } else {
+          console.error("Error deleting product");
+        }
+      } catch (error) {
+        console.error("Error deleting product:", error);
       }
-    } catch (error) {
-      console.error("Error deleting product:", error);
     }
   };
